Added working background service generating transaction daata

+++ src/TransactionSimulator/TransactionSimulator.Models/Card.cs
@@ -4,6 +4,6 @@ public class Card
 {
     public int Id { get; set; }
     public int UserId { get; set; }
-    public float CardLimit { get; set; }
-    public float MaxLimit { get; set; }
+    public decimal CardLimit { get; set; }
+    public decimal MaxLimit { get; set; }
 }
