+++ src/TransactionSimulator/TransactionSimulator.Repositories/Interfaces/ICardRepository.cs
@@ -7,5 +7,5 @@ public interface ICardRepository
     void AddCard(Card card);
     Card GetCard(int id);
     IList<int> GetCardIds();
-    void SubtractMoney(int id, float amount);
+    void SubtractMoney(int id, decimal amount);
 }
